+++ CHANGELOG.md
@@ -4,6 +4,26 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 0.4.1
+### Added
+- Added API POST lookup/shellsByAssetLink to retrieve shell ids without base64 encryption.
+- Added v3.1.0 postman collection for new API.
+## fixed
+- Changed API path from api/v3.0/ to api/v3/
+
+## 0.4.1-RC2
+### Added
+- Update Springboot to version 3.2.3
+- Update postgres dependency version to 42.7.2
+
+## fixed
+
+## 0.4.1-RC1
+### Added
+- Granular access control
+
+## fixed
+
 ## 0.3.23
 ### Added